requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain_huggingface
sentence_transformers
faiss-cpu
langchain_groq
streamlit
lxml
langchain_chroma
arxiv
wikipedia
langchainhub
duckduckgo-search